Removed string literals

+++ Source/PluginProcessor.cpp
@@ -96,7 +96,7 @@ void Fader_RiderAudioProcessor::changeProgramName(int /*index*/, const juce::Str
 //==============================================================================
 void Fader_RiderAudioProcessor::prepareToPlay(double sampleRate, int samplesPerBlock)
 {
-	m_pValueTreeState->SetSampleRate(sampleRate);
+	m_pValueTreeState->SetSampleRate(static_cast<int>(sampleRate));
 	m_pValueTreeState->SetSamplesPerBlock(samplesPerBlock);
 	m_pValueTreeState->UpdateParameterSettings();
 	//m_pValueTreeState->SetBlockRate(static_cast<float>(sampleRate / samplesPerBlock));
